Refactor Touchpad
